rename HasMiddleware
