+++ src/Piston.php
@@ -8,7 +8,7 @@ use League\Container\ServiceProvider;
 use League\Route\RouteCollection;
 use Psr\Http\Message\RequestInterface;
 use Refinery29\Piston\Middleware\HasMiddleware;
-use Refinery29\Piston\Middleware\HasPipeline;
+use Refinery29\Piston\Middleware\HasMiddlewareTrait;
 use Refinery29\Piston\Middleware\PipelineProcessor;
 use Refinery29\Piston\Middleware\Request\RequestPipeline;
 use Refinery29\Piston\Middleware\Subject;
@@ -17,9 +17,9 @@ use Refinery29\Piston\Router\RouteGroup;
 use Zend\Diactoros\Response\EmitterInterface;
 use Zend\Diactoros\Response\SapiEmitter;
 
-final class Piston extends RouteCollection implements HasPipeline
+final class Piston extends RouteCollection implements HasMiddleware
 {
-    use HasMiddleware;
+    use HasMiddlewareTrait;
 
     /**
      * @var Request
